% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/namespace_id_addin.R
\name{namespace_id_addin}
\alias{namespace_id_addin}
\title{Wrap an ID of a Shiny input in a call to `ns("")`}
\usage{
namespace_id_addin()
}
\description{
Call this addin after selecting the ID of an input in your source code. The
addin automatically adds quotes so do not select the quotes. The functionality
is currently extremely limited and only works for a very narrow usecase
(because this is all I needed and wanted to get it done in 5 minutes).
}
